backend/requirements.txt
flask
opencv-python
python-dotenv
requests
flask-cors
gunicorn
flask-wtf
pillow
flask-sqlalchemy
google-cloud-vision

